[c] - Move date utils to lib
